Added in drafts of Mongo and Dynamo services/facades.  API can now bulk insert from file path or multi-part MIME file stream (but nothing else). Lots of tests and error handling still needed.

+++ Apteco.PullMarketing/Models/Records/BulkUpsertRecordResults.cs
@@ -17,12 +17,12 @@ namespace Apteco.PullMarketing.Models.Records
     /// The number of records upserted
     /// </summary>
     [Required]
-    public int NumberOfRecordsUpserted { get; set; }
+    public long NumberOfRecordsUpserted { get; set; }
 
     /// <summary>
     /// The number of records skipped
     /// </summary>
     [Required]
-    public int NumberOfRecordsSkipped{ get; set; }
+    public long NumberOfRecordsSkipped { get; set; }
   }
 }
